totally simplify and refactor notifications
