+++ src/Illuminate/Notifications/Events/NotificationSent.php
@@ -2,25 +2,34 @@
 
 namespace Illuminate\Notifications\Events;
 
-use Illuminate\Notifications\Channels\Notification;
+use Illuminate\Notifications\Notification;
 
 class NotificationSent
 {
     /**
+     * The notifiable entity who received the notification.
+     *
+     * @var mixed
+     */
+    public $notifiable;
+
+    /**
      * The notification instance.
      *
-     * @var \Illuminate\Notifications\Channels\Notification
+     * @var \Illuminate\Notifications\Notification
      */
     public $notification;
 
     /**
      * Create a new event instance.
      *
-     * @param  \Illuminate\Notifications\Channels\Notification  $notification
+     * @param  mixed  $notifiable
+     * @param  \Illuminate\Notifications\Notification  $notification
      * @return void
      */
-    public function __construct(Notification $notification)
+    public function __construct($notifiable, $notification)
     {
+        $this->notifiable = $notifiable;
         $this->notification = $notification;
     }
 }
