+++ tests/Notifications/NotificationNotificationTest.php
@@ -1,37 +1,22 @@
 <?php
 
 use Illuminate\Notifications\Notification;
-use Illuminate\Notifications\Channels\Notification as ChannelNotification;
 
 class NotificationNotificationTest extends PHPUnit_Framework_TestCase
 {
     public function testLevelCanBeRetrieved()
     {
         $notification = new Notification;
-        $this->assertEquals('info', $notification->level());
+        $this->assertEquals('info', $notification->level);
 
         $notification = new NotificationTestNotification;
-        $this->assertEquals('error', $notification->level());
-    }
-
-    public function testSubjectCanBeRetrieved()
-    {
-        $notification = new NotificationTestNotification;
-        $this->assertEquals('Notification Test Notification', $notification->subject());
-
-        $notification = new NotificationTestNotificationWithSubject;
-        $this->assertEquals('Zonda', $notification->subject());
-    }
-
-    public function testMessageBuilderCanBeRetrieved()
-    {
-        $notification = new Notification;
-        $this->assertInstanceOf('Illuminate\Notifications\MessageBuilder', $notification->line('Something'));
+        $notification->level('error');
+        $this->assertEquals('error', $notification->level);
     }
 
     public function testChannelNotificationFormatsMultiLineText()
     {
-        $notification = new ChannelNotification([]);
+        $notification = new Notification([]);
         $notification->with('
             This is a
             single line of text.
@@ -39,7 +24,7 @@ class NotificationNotificationTest extends PHPUnit_Framework_TestCase
 
         $this->assertEquals('This is a single line of text.', $notification->introLines[0]);
 
-        $notification = new ChannelNotification([]);
+        $notification = new Notification([]);
         $notification->with([
             'This is a',
             'single line of text.',
